saltproc is a class now
